On the first database write after a flush Gap4 will check the global time stamp in the Aux header to make sure that the disk copy matches it's own cached in-mory copy.
The rationale behind this is that users still manually remove the BUSY
file because "they know best". They then act all suprised when their
database becomes corrupt. Our alternative here is to check the master
time-stamp in the Aux header and to simply abort Gap4 (losing all unsaved
edits) if it has been changed external to this process.

This still leaves room for race conditions (multiple people editing
simultaenously), but it covers the more common case of a gap4 session
left open for ages and then going back to it later.


git-svn-id: svn://svn.code.sf.net/p/staden/code/staden/trunk@255 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1