Decreased the impact of primer3 scores on the 'cost_mult' parameter. It's now *= 1+score/50 instead of *= 1+score/20.
When -chk_template_stat == 0 we now set the template ranges to be the
observed location of vector +/- 1Kb.

"Solution not adjoining problem end" is now multiplicative in cost
rather than additive.

Tweaks for start location of where to search for primers when we're
trying to solve a + strand problem from the right end (ie the left end
failed) and where start != end. We back up from the requested pos by
half a read length.


git-svn-id: svn://svn.code.sf.net/p/staden/code/staden/trunk@366 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1