Editor sequence-name panel now remembers the X scrollbar position between contig editor uses (within a single session).
Experimental "tag popup" system. Currently only suitable for the
inhouse ExoSeq project and so disabled in the default gaprc.

"Show template names" option to toggle whether reading names or
template names are shown in the left name panel.

The "Join to" function now attempts to automatically complement a
contig if required (based on read-pair primer types).


git-svn-id: svn://svn.code.sf.net/p/staden/code/staden/trunk@445 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1