Added destroy_malign_counts, malign_insert_scores and malign_recalc_scores functions. The latter of these allow for updating of the malign scores and consensus elements after a contigl list has been updated. Used as a fast alternative to rebuilding from scratch.
Removed unused msegs element of MALIGN struct.

Allow for gap_open and gap_extend to be passed in when building an
MALIGN. These are then used instead of the previous hardwired settings.

Fixed a bug in affine_malign where the first rows of the F and G
matrix were undefined meaning that the computation of G (gap in Y)
used uninitialised data. This was the cause of sometimes giving
daft alignments. The same bug probably exists in all the other
alignment functions here, but has not been checked or fixed.


git-svn-id: svn://svn.code.sf.net/p/staden/code/staden/trunk@787 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1