This attempts to reorder an ACE file so that the reads within it are sorted in left to right order.
The purpose for this is to improve efficiency when converting to a Gap5
database. While it'll work regardless of the read order (provided the
reads are presented in the same order as listed int he AF lines), it is
inefficient if the reads for a small region are spread throughout a large
area of disk.


git-svn-id: svn://svn.code.sf.net/p/staden/code/staden/trunk@1317 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1