Changed the format of the on-disc GT_Seq, GT_Bin and GT_Range structures. NB: *INCOMPATIBLE* file format with previous release.
Sequences no longer hold their paired sequence record directly in the
sequence struct. Instead it's in the bin range array. To facilitate
fast lookup the sequence holds the index in the bin range array
referring back to that sequence.

This allowed a change to how the pairing is computed too. In
contigs_seqs_in_range we hash on record number and so link each
sequence directly to its pair. This means that we don't need to load
any sequence structs at all when plotting graphical displays like the
template display, except for the rair cases at the edge of the window
being plotted. (The I/O improvement in number of seeks and reads is
somtimes a 10 fold drop.)

New function sequence_get_pair() hides away the logic to compute the
other end of a sequence pair.


git-svn-id: svn://svn.code.sf.net/p/staden/code/staden/trunk@1472 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1