Changes to the template display following the sequence and bin/range structure changes.
We now have an accurate and approximate mode. Both need MUCH less I/O
than before, but approximate mode saves further I/O by not bothering
to work out whether the other end of a template, which is not observed
on screen, is in this contig and if so whether it is to the right or
left of the sequence we did observe.


git-svn-id: svn://svn.code.sf.net/p/staden/code/staden/trunk@1475 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1