Added code (currently "#if 0"ed out) to attempt to cache the consensus in tracks per bin. Specifically the plan was to store the A,C,G,T,* accumulated log(probability) values for data held within that bin as a track. It means we can compute the consensus without having to load all the invdividual sequences.
In practice, it does speed up the consensus generation when cached
(usually only barely, but occasionally by > 10x), but the downsides
are considerable. It needs read-write access, and worst of all the
space it takes up is very substantial - often far more than the rest
of the database. The existing implementation also is only approximate
as we store the floating point values in very low precision.

Given the size problems I'm planning on reimplementing this via
another means, but the code is checked in here (disabled) so I can
revert or refer back to it at some stage if required.

Also see the code changes to tg_iface_g.c.


git-svn-id: svn://svn.code.sf.net/p/staden/code/staden/trunk@1832 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1