Fixed bug in consensus_valid_range where a right hand cutoff of > 1Kb was reported as 1Kb only.
Fixed a bug with using cached consensus data from bins that don't
quite butt up to one another. Why this happens I am unsure still
(likely a bug in contig joining), but the defensive approach is to
detect it and fall back to filling in the consensus gaps using the old
technique.

Also fixed a consensus algorithm bug where we could overflow the cache
and break things. We use cache_incr/decr now to protect against this.


git-svn-id: svn://svn.code.sf.net/p/staden/code/staden/trunk@1909 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1