Fixed display and undo issues relating to moving sequences, especially moving them left from the start of the contig into negative coorindates.
We're now more careful about caching of data which may change, such as
the contig struct and the root bin it points to. In Tcl we resolve
this largely by not caching the object, but only the record number. In
C we resolved it by implementing a manual way to repopulate cached
contig struct on demand (the decr_contig and incr_contig editor
methods), although it may be preferable to replace these by record
number caching instead at some stage.


git-svn-id: svn://svn.code.sf.net/p/staden/code/staden/trunk@2006 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1