We're now more explicit about our library dependencies. This should aid compilation on MacOS X and Windows (untested).
Also made sure that for TCL only we add $(TCL_LIB) and for Tcl/Tk both
we use only $(TK_LIB) and not both. This is because adding @TCL_LIBS@
and @TK_LIBS@ to the system.mk.in has lead to some dependent libraries
being specified twice if we explicitly ask to use both $(TCL_LIB) and
$(TK_LIB). For libieee.a this causes problems due to a global version
symbol, causing a link error.


git-svn-id: svn://svn.code.sf.net/p/staden/code/staden/trunk@2110 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1