Fixed our check for bin->{start_used,start_end} being valid from comparing whether they are equal to checking whether bin->rng is blank.
The former method meant that a bin consisting solely of a 1 base tag
(eg start_used=1000, end_used=1000) was considered empty. The next
time we add another tag to the bin, start/end weren't extended but
instead moved as-if initialising an empty bin. This caused tags to be
invisible in some corner cases.


git-svn-id: svn://svn.code.sf.net/p/staden/code/staden/trunk@2179 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1