Updated the compute_ypos function in tg_contig.c (note, not the one in template_display.c) to keep track of previous Y coordinates.
We achieve this by modifying the low-level cached bin-range structures
to also include a Y (in memory and not written to disk). If our next
query already has Y coordinates used then we prefer these over
allocating new ones.

This prevents the "rotating" effect of short-read data when scrolling
right in the contig editor, although at the moment this still happens
when scrolling left.


git-svn-id: svn://svn.code.sf.net/p/staden/code/staden/trunk@2184 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1