Added extra paranoia after observing a corrupted database, where another record type got written to aux slot 0 (should be a GDatabase record), despite an apparent check for this with an assert.
The cached_item struct holds both a record and a View; a handle to
that record. The View in turn has a cache of the aux data for this
record, which in turn holds the record number again. Double check that
these two are identical so at least we die rather than corrupt (and
hopefully help to point out what went wrong if so).



git-svn-id: svn://svn.code.sf.net/p/staden/code/staden/trunk@2396 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1