Added handling for cigar N (reference skip) alignments.
For now we just create a separate sequence, but we need a more general
purpose mechanism of dealing with templates with more than 2
sequences.

TODO: Check the TC field in sam aux as an indicator of whether it's
acceptable to remove the hashed template name when we see the second
record. Also set this mode when finding a cigar N op.


git-svn-id: svn://svn.code.sf.net/p/staden/code/staden/trunk@2447 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1