Implemented a new consensus algorithm to compute probabilities for both pure single-base columns of data and mixed heterozygous columns.
This means it can call the consensus as likely consisting of, say, an
AT mix instead of pure A or pure T. We still also call the most likely
homozygous consensus, but if it is more likely to be a heterozygous
mix then the probability assigned to a homozygous call is therefore
far lower than before.

My belief is that this will prevent major conflicts still being given
very high homozygous consensus values simply because one base type
outweighs another. This work also forms the basis of SNP calling or
misassembly detection.


git-svn-id: svn://svn.code.sf.net/p/staden/code/staden/trunk@2500 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1