Bug fix to the template pairing.
The recent change to add the quality plot changed how the detection of
"done this pair" code worked, from zeroing record number to setting a
flag (as we needed the record left intact). Unfortunately it was
bugged. The impact was that using filters would lose many pairs and
they'd stay lost until we had to query a new range.


git-svn-id: svn://svn.code.sf.net/p/staden/code/staden/trunk@2524 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1