Removed a reference count leak in the contig iterator (leaking bins at times).
Made contig_destroy() decrement the size of the contig_order array.
This wasn't strictly a problem as we always use io->db.Ncontigs to
work out how many contigs there are, but it's inefficient (writing out
more data than we need to) and the new check_database code just
happened to notice it due to extra-anal checking.


git-svn-id: svn://svn.code.sf.net/p/staden/code/staden/trunk@2559 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1