Fixed the usage of obj_rec in anno_ele_t. Obj_type/obj_rec (and pair_rec in range_t) point to the record that is being annotated. This works well for sequences, but is problematic for contigs. Eg consider breaking a contig in two; all consensus annotations in the right hand half will need their obj_rec (and pair_rec) fields changing.
Mostly this worked as we simply didn't bother and instead took
consensus annotations to be on the contig they are within, regardless
of the contents of obj_rec or pair_rec, but not in all cases. Now we
force obj_rec to be zero when obj_type is GT_Contig and fixed the
exceptions where we failed to ignore the field.


git-svn-id: svn://svn.code.sf.net/p/staden/code/staden/trunk@2572 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1