Fix to severe bug in shuffle pads.
In rare cases it could introduce a column of pads itself, and then
remove it again in the final step. If during this stage it didn't make
any other edits then it did not save the data and so the column being
removed was incorrect - effectively deleting a real column of bases.

This has now been fixed, but also added a contig_delete_pad function
which acts as contig_delete_base but with additional checks to ensure
it doesn't accidentally delete something else.


git-svn-id: svn://svn.code.sf.net/p/staden/code/staden/trunk@2736 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1