Scary change, but I *think* a correct one... It seems to pass my gap5_thrash tests anyway.
When doing cache_rw on an object from a derived (aka child) io, we use
copy-on-write methods to duplicate the object from the parent io into
the child io. The address we return to the caller is now the child io
object instead. This is implemented via cache_dup(). This is used
within the contig editor to allow for edits to be not live and also to
allow exit without saving.

When incremementing or decrementing the reference count on a child io
object we were previously just working on that object itself. However
the cache_dup function didn't copy the full reference count from the
parent - this is perhaps a bug. Consequentially it's possible to
perform several cache_incr on a read-only object, call cache_rw to get
a new one (now with ref count 1 again), and then cache_decr on the
read-write object and trigger an error due to negative reference
counts.

Rather than attempt to fix this via copying the master ref count
into the child io object, which itself causes other errors, I now do
all reference counting on the master object instead. Hence a
cache_incr or cache_decr on an object in the child io will update the
ref count in the parent io. (This is perhaps not implemented in the
most efficient manner, using a separate cache_search, but if it has
performance issues consider implementing a master pointer to
cached_item struct.)

I believe this is robust given the child io only contains cache_rw
objects and we'll never be wanting to purge from the child - making it
more hash than hache.



git-svn-id: svn://svn.code.sf.net/p/staden/code/staden/trunk@2810 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1