Warning: I'm not *100%* certain this is correct!
Swapped the unlock and commit calls during the flush procedure.
Previously this was causing B+Tree writes to be auto-flushed giving
lots of index updates. Now they'll also be appended on the "to flush"
list due to the lock and updated in one step along with the other
records recently written. The theory was to improve performance, but
I'm not sure it makes much difference.



git-svn-id: svn://svn.code.sf.net/p/staden/code/staden/trunk@2907 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1