Bug fixes to the ACE format reading code.
It now handles (ie ignores instead of choking) WR{} records.

It also now spots left/right clip points as both -1 and discards such
entries and unaligned data.

Finally it handles the lack of DS lines. Previously it added reads
when seeing a DS line. Now it adds reads when it sees either a new CO
or RD line so the lack of DS lines does not mean a total lack of
imported data.


git-svn-id: svn://svn.code.sf.net/p/staden/code/staden/trunk@2919 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1