Added KP_Enter binding for position panel (in addition to existing Return key binding).
Changed Delete key to delete under the cursor and not move, vs remove
from left of cursor and move for Backspace.

Added functionality to move pads left and right with Control-L and
Control-R.  For those that like to live dangerously, bases can also be
moved if you use Alt-L and Alt-R bindings.


git-svn-id: svn://svn.code.sf.net/p/staden/code/staden/trunk@2942 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1