Found a cause of consensus caching to sometimes use different bins, causing unnecessary work.
Even when we have a fully cached consensus, updated by doing "Save
consensus" for all contigs, asking for smaller ranges of consensus (eg
search by consensus quality) would sometimes return a different set of
bins to hold the cached consensus, causing recomputation of the
consensus again.

The cause is that the function returns the first (ie largest) bins
within a range smaller than a fixed size (100Kb). If we have a bin
with only one child node to the right, then on a full range query that
child node may be returned, but on a smaller range query it may be
excluded and the parent node is instead returned.

The code now explictly checks for this case and filters such bins.


git-svn-id: svn://svn.code.sf.net/p/staden/code/staden/trunk@2947 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1