Fixed analysis of template status.
For the contig editor, via sequence_get_template_info(), we now
consider TEMPLATE_ORIENT as a valid return status.

For the template display, fixed a bug where assumptions were that
r1->comp and r2->comp would be identical. This meant templates
spanning bins could sometimes be considered in the wrong
orientation. We now correct the COMP1 and COMP2 flags taking this into
account.


git-svn-id: svn://svn.code.sf.net/p/staden/code/staden/trunk@2963 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1