Major overhaul of GFF import and export in Gap5.
(Some of this changes code in Gap4 as the tag database and ACD parsing
is shared between the two, but in an invisible manner.)

We now have a GUI for GFF tags using the #!acdtag syntax. The ACD tag
code was extended slightly to allow querying of ACD variables so the
unhandled key=value attribute pairs in GFF can be placed inside a
text box.

Also moved the various escaping and unescaping functions out from
export_contigs.c to Misc/strings.c as it's now used in multiple places
- all in gap5 atm, but Misc is the right spot.



git-svn-id: svn://svn.code.sf.net/p/staden/code/staden/trunk@3017 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1