Updates to the copy-on-write caching layer to allow for the possibility of child IOs of child IOs (ie grandchild IOs).
In practice it's not finished, but it doesn't harm normal operation
either and the commenting has been improved, with a few extra asserts
too. So we'll keep this variant.

There is also a new gio_base(io) function to return the base io rather
than just using io->base ? io->base : io.


git-svn-id: svn://svn.code.sf.net/p/staden/code/staden/trunk@3058 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1