Added an option to perform bulk contig renaming.
This operates either on a list or on a pattern to check against all
contig names.  The replacement text has a single printf rule to
replace %d by an automatically iterated contig number. Eg replacing
everything with "Contig_%06d" to get Contig_000001 upwards.

If operating in pattern mode it can use filename glob style wildcards
("Node*Contig*"). In this case the replacement pattern can also use
\1, \2 etc regexp matches to utilise whatever the * matched. This
allows for renaming of "Contig*" to "C\1". (Should this be "C*" too
for ease?)


git-svn-id: svn://svn.code.sf.net/p/staden/code/staden/trunk@3068 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1