Added user interface to import_reads and to make an empty database.
Added menu items and interfaces to allow users to import reads without
having to use tg_index.  Also allow them to create an empty database.

Fixed bitmap values for the main menus so that items are enabled and disabled
correctly as the state changes from started to database open but empty to
open with data in it.

Prevented divide by zero in DatabaseInfo when run on an empty database.


git-svn-id: svn://svn.code.sf.net/p/staden/code/staden/trunk@3157 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1