More rigorous calculation of invalid read-pairs.
We ensure contig-spanning pairs have an sequence pointing towards the
problem rather than away from it, and also that it's not close to the
contig end and pointing off it as it may then be a real join.

Singletons for reads from paired end libraries now also get their own
scoring category (currently set to 0 while I test they're worth).


git-svn-id: svn://svn.code.sf.net/p/staden/code/staden/trunk@3440 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1