Fixed reference count on libraries in auto break.
compute_lib_type was not calling cache_incr on the library_t pointers
that it stored in a hash table.  This meant the pointers could go stale
if the library information later got pushed out of the cache, leading to
a segmentation fault.

compute_lib_type now calls cache_incr on the library_t pointer to prevent this.
confirm_gaps calls cache_decr on all of the items in the hash table when it
has finished with them.



git-svn-id: svn://svn.code.sf.net/p/staden/code/staden/trunk@3442 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1