Corrected check_overlap_pairs() to only count failed pairs in the overlap region specified.
This isn't ideal, but at least it's the same count used for good
pairs, so it doesn't artificially think the worst about all joins.

Ideally we should switch to the pair analysis used in auto-break, on a
simulated joined up contig. (Or maybe just make the join in a child io,
evaluate it, and then destroy the io). 


git-svn-id: svn://svn.code.sf.net/p/staden/code/staden/trunk@3483 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1