Added an option to Shuffle Pads to use the cutoff/soft-clip data.
This is quite complex as it performs multiple passes. Firstly it looks
for "coherent soft-clips"; regions with multiple sequences
soft-clipped with the cutoffs in agreement with one another. Then it
extends sequences into the cutoffs before running shuffle pads.

Once realigned, it then inspects the data to check that the unclipped
reads align. If so then they are kept, if not then they are tagged
with CLIP and the clips are trimmed back to where they agree with the
consensus.

Finally special care is taken around consensus Ns. These are included
in local conventions by faking scaffolds, placing contigs next to each
other in the same "contig". Sometimes the realignment will resolve and
join them together, but often it produces unaligned data. In that case
we attempt to unclip and recreate the N as a internal scaffold marker,
but it's not bullet proof.


git-svn-id: svn://svn.code.sf.net/p/staden/code/staden/trunk@3583 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1