Disabled fix for annotations that do not overlap the annotated object.
Trying to move the annotation to a new bin causes problems in bin_walk which
expects the items in the bins to remain stable.  This leads it to try to
fix problems that aren't really there resulting in a database that is in
a much worse state that it was to start with.  Disabling the annotation
fix prevents this.


git-svn-id: svn://svn.code.sf.net/p/staden/code/staden/trunk@3600 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1