Made bin_add_to_range explicitly check for sequences beyond clipped_start/end
bin_add_to_range would fail to set (*c)->clipped_timestamp = 0 if a sequence
being added was within the existing contig used start/end but caused the
clipped_start/end to change.  Added a check to see if the new range may
go beyond clipped_start or clipped_end, and reset clipped_timestamp if
this is the case.


git-svn-id: svn://svn.code.sf.net/p/staden/code/staden/trunk@3622 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1